Porting test code to AutoMATES

+++ tests/data/program_analysis/array_func_loop/array-to-func_06.py
@@ -1,23 +1,23 @@
 import sys
+import os
 from typing import List
 import math
-from program_analysis.for2py.format import *
-from program_analysis.for2py.arrays import *
-from program_analysis.for2py.static_save import *
-from program_analysis.for2py.strings import *
+from delphi.translators.for2py.format import *
+from delphi.translators.for2py.arrays import *
+from delphi.translators.for2py.static_save import *
+from delphi.translators.for2py.strings import *
+from delphi.translators.for2py import intrinsics
 from dataclasses import dataclass
-from program_analysis.for2py.types_ext import Float32
-import program_analysis.for2py.math_ext as math
+from delphi.translators.for2py.types_ext import Float32
+import delphi.translators.for2py.math_ext as math
 from numbers import Real
 from random import random
 
 
 def update_mean_var(means: Array, vars: Array, k: List[int], n: List[int], runs: List[int]):
     tmax: List[int] = [100]
-    
-    
-    means.set_((k[0]), (means.get_((k[0])) + ((n[0] - means.get_((k[0]))) / (runs[0] + 1))))
-    vars.set_((k[0]), (vars.get_((k[0])) + (((runs[0] / (runs[0] + 1)) * (n[0] - means.get_((k[0])))) * (n[0] - means.get_((k[0]))))))
+    means.set_((k[0]), Float32((means.get_((k[0])) + ((n[0] - means.get_((k[0]))) / (runs[0] + 1)))))
+    vars.set_((k[0]), Float32((vars.get_((k[0])) + (((runs[0] / (runs[0] + 1)) * (n[0] - means.get_((k[0])))) * (n[0] - means.get_((k[0])))))))
     
 
 def gillespie(means: Array, vars: Array):
@@ -26,11 +26,9 @@ def gillespie(means: Array, vars: Array):
     sample: List[int] = [None]
     i: List[int] = [None]
     j: List[int] = [None]
-    
-    
     samples = Array(float, [(0, tmax[0])])
     for i[0] in range(0, tmax[0]+1):
-        samples.set_((i[0]), i[0])
+        samples.set_((i[0]), Float32(i[0]))
     for runs[0] in range(0, tmax[0]+1):
         j[0] = 0
         while (j[0] <= tmax[0]):
